pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ubunturouter"
version = "0.1.0"
description = "Ubuntu-based modern software router framework"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "python-pam>=2.0",
    "pyjwt>=2.8",
    "docker>=6.0",
    "paramiko>=3.0",
    "aiohttp>=3.9",
]

[project.scripts]
urctl = "ubunturouter.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
